Validate new index of moved board
Emits BOARD_MOVED from workspaceController. Attaches a BOARD_MOVED
listener to boardList that checks if the board ID and index in the
workspace model match the ID and index of the moved board list item. If
both match, the move is stopped, leaving the item in its current
position. If either don't match, the move is canceled, returning the
item to its initial index, and an error is thrown.